+++ scripts/macros.js
@@ -130,6 +130,7 @@ export {spiritualWeapon} from './macros/2024/spells/spiritualWeapon.js';
 export {staggeringSmite} from './macros/2024/spells/staggeringSmite.js';
 export {summonAberration, summonAberrationWhisperingAura} from './macros/2024/spells/summonAberration.js';
 export {summonBeast} from './macros/2024/spells/summonBeast.js';
+export {summonCelestial, summonCelestialMace} from './macros/2024/spells/summonCelestial.js';
 export {synapticStatic} from './macros/2024/spells/synapticStatic.js';
 export {teleport} from './macros/2024/spells/teleport.js';
 export {thornWhip} from './macros/2024/spells/thornWhip.js';